+++ pkg/test/setup/setup.go
@@ -57,8 +57,10 @@ func CommonWithSchemaLoaders(schemaLoaders []SchemaLoader, flags ...string) (str
 	addr := fmt.Sprintf("%s:%d", host, ports[0])
 	httpAddr := fmt.Sprintf("%s:%d", host, ports[1])
 	ff := []string{
-		"--addr=" + addr,
-		"--http-addr=" + httpAddr,
+		"--grpc-host=" + host,
+		fmt.Sprintf("--grpc-port=%d", ports[0]),
+		"--http-host=" + host,
+		fmt.Sprintf("--http-port=%d", ports[1]),
 		"--http-grpc-addr=" + addr,
 		"--stream-root-path=" + path,
 		"--measure-root-path=" + path,
@@ -92,7 +94,7 @@ func modules(schemaLoaders []SchemaLoader, flags []string) func() {
 	q, err := query.NewService(context.TODO(), streamSvc, measureSvc, metaSvc, pipeline)
 	gomega.Expect(err).NotTo(gomega.HaveOccurred())
 	tcp := grpc.NewServer(context.TODO(), pipeline, metaSvc)
-	httpServer := http.NewService()
+	httpServer := http.NewServer()
 
 	units := []run.Unit{
 		pipeline,
@@ -129,11 +131,11 @@ func (p *preloadService) Name() string {
 	return "preload-" + p.name
 }
 
-func (p *preloadService) PreRun() error {
+func (p *preloadService) PreRun(ctx context.Context) error {
 	if p.name == "stream" {
-		return test_stream.PreloadSchema(p.metaSvc.SchemaRegistry())
+		return test_stream.PreloadSchema(ctx, p.metaSvc.SchemaRegistry())
 	}
-	return test_measure.PreloadSchema(p.metaSvc.SchemaRegistry())
+	return test_measure.PreloadSchema(ctx, p.metaSvc.SchemaRegistry())
 }
 
 func (p *preloadService) SetMeta(meta metadata.Service) {